+++ styles.js
@@ -3,7 +3,5 @@ export * as gnome from './palette/gnome.js';
 export { btn } from './button.js';
 export { reset } from './reset.js';
 export { sheetToFile, sheetToLink } from './utils.js';
-export {
-	baseTheme, darkTheme, lightTheme, componentBase, componentDarkTheme,
-	componentLightTheme,
-} from  './theme.js';
+export * from './properties.js';
+export * from  './theme.js';